# dedup extra handshakes to speedup cracking time. …