This will make less work by removing extra salts.

```
$ ./a.out test.hashcat foo
Read 966 handshakes
Filtered: 835 handshakes
Wrote: 131 handshakes
```

```
$ ./a.out foo foo
Read 131 handshakes
Filtered: 0 handshakes
Wrote: 131 handshakes
```