@ChrisLundquist 2018-01-14
speed up in practice:

Before
```
Session..........: hashcat
Status...........: Quit
Hash.Type........: WPA/WPA2
Hash.Target......: .\total.hashcat
Time.Started.....: Sat Jan 13 22:58:13 2018 (26 secs)
Time.Estimated...: Mon Jan 15 01:30:23 2018 (1 day, 2 hours)
Guess.Base.......: File (.\rockyou.txt)
Guess.Mod........: Rules (.\rules\best64.rule)
Guess.Queue......: 1/1 (100.00%)
Speed.Dev.#1.....:   373.7 kH/s (6.51ms)
Speed.Dev.#2.....:   373.8 kH/s (6.63ms)
Speed.Dev.#*.....:   747.5 kH/s
Recovered........: 1/535 (0.19%) Digests, 1/67 (1.49%) Salts
Progress.........: 1530568060/74002677056 (2.07%)
Rejected.........: 1511071100/1530568060 (98.73%)
Restore.Point....: 0/14344384 (0.00%)
Candidates.#1....: 1234567893 -> classica3
Candidates.#2....: 021119933 -> LIMERICK3
HWMon.Dev.#1.....: Temp: 54c Fan: 33% Util:100% Core:1987MHz Mem:5005MHz Bus:16
HWMon.Dev.#2.....: Temp: 55c Fan: 33% Util:100% Core:1974MHz Mem:5005MHz Bus:16
```

After:
```
Session..........: hashcat
Status...........: Running
Hash.Type........: WPA/WPA2
Hash.Target......: .\total.hashcat
Time.Started.....: Sun Jan 14 11:21:18 2018 (5 mins, 35 secs)
Time.Estimated...: Mon Jan 15 07:54:38 2018 (20 hours, 27 mins)
Guess.Base.......: File (.\rockyou.txt)
Guess.Mod........: Rules (.\rules\best64.rule)
Guess.Queue......: 1/1 (100.00%)
Speed.Dev.#1.....:   505.7 kH/s (6.34ms)
Speed.Dev.#2.....:   503.1 kH/s (6.36ms)
Speed.Dev.#*.....:  1008.9 kH/s
Recovered........: 2/71 (2.82%) Digests, 2/71 (2.82%) Salts
Progress.........: 1975511244/78420747328 (2.52%)
Rejected.........: 1601284300/1975511244 (81.06%)
Restore.Point....: 0/14344384 (0.00%)
Candidates.#1....: 0211199312 -> downpour12
Candidates.#2....: 1234567897 -> classica7
HWMon.Dev.#1.....: Temp: 75c Fan: 43% Util: 99% Core:1885MHz Mem:5005MHz Bus:16
HWMon.Dev.#2.....: Temp: 77c Fan: 64% Util: 99% Core:1885MHz Mem:5005MHz Bus:16
```

It appears fewer salts helps keep the GPUs busier and we don't lose hash rate.