@ZerBea 2018-01-15
Possible that you are going in the wrong direction. Stripping by ESSID and message_pair is not a good idea and you will lose many crackable handshakes and grep uncrackables instead.

Choosing a good mix of handshakes and hashcat parameters is a better way:
$ hashcat -m 2500 --benchmark
hashcat (v4.0.1-78-g185104b5) starting in benchmark mode...
Device #1: GeForce GTX 1080 Ti, 2792/11169 MB allocatable, 28MCU
Speed.Dev.#1.....:   540.8 kH/s (52.79ms)

real life eaxmple 1:
Speed.Dev.#1.....:   534.5 kH/s (6.59ms)
Recovered........: 0/12885 (0.00%) Digests, 0/8523 (0.00%) Salts

real life eaxmple 2:
Speed.Dev.#1.....:   516.4 kH/s (6.60ms)
Recovered........: 2140/1094175 (0.20%) Digests, 200/156594 (0.13%) Salts