@jsteube 2018-01-15
I agree to @ZerBea because the use multiple handshakes is done on purpose. It's exactly what aircrack & co. is missing and it's one of the reason we started this extra format. 

Maybe check out the --nonce-error-corrections 0 to the hashcat commandline, it can have a big performance impact on hccapx files with many handshakes. Of course make sure you have a clean recording before you use it in real life.