@ChrisLundquist 2018-01-15
Hmm, I'd love to learn more here.
The above code keeps one of each type of message pair for every network.

We can see that hashcat disregards some portion of these since it only cares about 71 digests out of the 131. Off the top of my head, it was trying crack message pairs that hand shakes 3+4 since that shows a complete login.

Is there currently an attack that leverages redundancy between multiple authentications for WPA? (WEP has this for sure)

In practice the above was my phone connecting to my wifi network as I came and went.
The only other way I've found to remove extraneous work, is to feed the pot into hashcat so it cracks each new capture's salt/nonce.

My goal is to understand what value we'd be losing by removing these duplicates.

As @ZerBea showed, 1080TIs should be around 500kH/s and 374kH/s seems like a notable drop.